Only merge to the latest revision per branch
In some cases the copyfrom data combined with the svn:mergeinfo data could
result in a lot of different revisions being merged.

For git we just need to let the different commits be reachable and in the
future git will be able to find the applicable change set.

Change how the parents of a commit are created so that if there are 10
revisions on a single branch there will only be 1 merged to (the most
recent revision).